requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
xgboost
Pillow
torch
torchvision
joblib